web/backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
SQLAlchemy==2.0.47
Flask-SQLAlchemy
typing_extensions==4.15.0
Werkzeug==3.1.6
gunicorn
psycopg2-binary
eventlet